src/main/java/com/example/demo/student/StudentService.java
package com.example.demo.student;

import com.example.demo.student.exception.BadRequestException;
import com.example.demo.student.exception.StudentNotFoundException;
import lombok.AllArgsConstructor;
import org.springframework.stereotype.Service;

import java.util.List;

@AllArgsConstructor
@Service
public class StudentService {

    private final StudentRepository studentRepository;

    public List<Student> getAllStudents() {
        return studentRepository.findAll();
    }

    public void addStudent(StudentRegistrationRequest request) {
        Boolean existsEmail = studentRepository
                .selectExistsEmail(request.email());

        if (existsEmail) {
            throw new BadRequestException(
                    "Email " + request.email() + " taken");
        }

        Student student = new Student(request.name(), request.email(), request.gender());

        studentRepository.save(student);
    }

    public void deleteStudent(Long studentId) {
        if(!studentRepository.existsById(studentId)) {
            throw new StudentNotFoundException(
                    "Student with id " + studentId + " does not exists");
        }
        studentRepository.deleteById(studentId);
    }
}
